+++ CHANGELOG.md
@@ -2,7 +2,12 @@
 
 ## Current (in progress)
 
+- Nothing yet
+
+## 3.3.1 (2022-01-11)
+
 - Fix fields empty value in admin form to allow for unsetting fields [#2691](https://github.com/opendatateam/udata/pull/2691)
+- :warning: Add a new required topic string field on reuses. The associated migration set default topic to `others` [#2689](https://github.com/opendatateam/udata/pull/2689)
 
 ## 3.3.0 (2021-12-10)